requirements.txt
fastapi
uvicorn
pydantic
google-genai
Pillow
python-multipart
